Avoid useless email if scheduled too often
Hotsync tries to sync as frequent as possible. If a sync has not ended
before the next one starts, simply exit, without sending any email.
In case of big amounts of data added in a short period, a single sync may require more
time, but it's not a problem. If a sync lasts more than two hours,
notify root.